// app/interfaces/MainComponent/DiscoverSection.ts
export interface IDiscoverSection {
    _id: string;
    id?: string;
    pagetitle?: string;
    pagedesctiption1?: string;
    pagedesctiption2?: string;
    title?: string;
    image?: string;
    description?: string;
}

export interface IDiscoverSectionInput {
    id?: string;
    pagetitle?: string;
    pagedesctiption1?: string;
    pagedesctiption2?: string;
    title?: string;
    image?: string;
    description?: string;
}